requirements.txt
flask
fastapi
pandas
numpy
pytesseract
pdf2image
Pillow
transformers
torch
streamlit
sqlite3
unittest
